Temporarily resolve "cannot pickle '_io.TextIOWrapper' object" error
This error was happening because the apply_async method (like all
python multiprocessing methods) has to "pickle" (aka serialize) the
object it is operating on. The Parser object can't be pickled because
it contains a filestream object, and filestream objects can't be
pickled.

To temporarily resolve this, I just removed multiprocessing from the
construct_commands method, so now the chunks are parsed sequentially
instead of in parallel, which is slower.

In the future, we should try to find a way (maybe a different library
for async code) that will work to make the parser faster. Or, if
that doesn't work, we can remove the break_file_into_chunks method,
since that is only needed for parallelizing the parser.