Remove EXAM_NUMBER env var
Comments and documentation said that we were using exam number in
evaluators and database tables, but we were not. It wasn't being used at
all, so I removed it. Eventually I think we'll need to add back in the
idea of an identifier for each exam/dataset, but we aren't there yet,
and this cruft was just confusing.