Update evaluator metadata instead of replacing records
Previously, when the import_evaluator_metadata command ran, it would
delete all existing evaluators and re-import them all as new records.
Now, it matches the records by name. If the name exists in the DB, it
updates the existing record, rather than creating a new one.

This is beneficial because when an EvaluatorMetadata record is deleted,
any records it is the parent of, such as EvaluatorResultSummary records,
would also be deleted, and we probably don't want that. Instead, we
would want the existing EvaluatorResultSummary records to be associated
with the updated EvaluatorMetadata record, which is what should happen
now.