More accurate names for settings files
1. helm_eks (formerly helm_dev) is what we'll use for all Alto
   environments (I hope -- if we can interpolate a few variables)
2. helm_local (formerly also helm_dev) is what we'll use when we're
   testing out the helm chart on a laptop

I deleted the file called `staging` because it didn't have a
sensible naming convention, and it didn't have any settings in it
anyway.

Now, when we run the helm chart locally, ./helm-install uses
values.yaml, which specifies the helm_local settings file.

When we run the helm chart from Jenkins, the build job calls
values-eks.yaml, which specifies the helm_eks settings file.