Use raw SQL query to associate previous records
The previous method to associate_previous_records pulled the entire list
of AccountActivity records into python memory, which crashes the process
when the number of records is really large.

The Django ORM does not have a way to update a table with the result of
a Window expression, so we had to resort to a raw SQL query to keep all
of the processing in the database and out of the python process.

This update makes the assocate_previous_records method significantly
faster.