WIP: Modify evaluator to save results in sql
Previously, saving eval results was failing when using a large dataset,
because the list of results was too large to store in python memory, and
the process would be killed.

Now, generate and save the results in SQL, without bringing them into
python memory. Unfortunately, saving field_values as a JSON value is too
complicated to be feasible when using raw SQL, so we have to modify how
we store and retrieve evaluator results.

TODO:
- update individual evaluators to return a queryset instead of a list
  of values
- update how we return eval results in the API
- update all tests affected by these changes