Vastly reduce number of database hits for acct summary
Previously, the code had to query the database ~8 times for EVERY
tradeline (so, 40+ queries for a result of 5 tradelines). Now, it
queries about once per tradeline.

Ideally this would make the same number of queries regardless of how
many tradelines are in the result, so this isn't perfect, but it's
better.