Improve efficiency for event summary view
Previously, this view would hit the database 5 times for each evaluator
that had a hit for this event, resulting in hundreds of total queries
(depending on how many evals there were). Now, this makes a the same
number of queries to load the page (4) regardless of how many evals hit.

This involved changing the serializer to be based on the
EvaluatorResultSummary model instead of the EvaluatorMetadata model.