Use utility methods for generating eval results
Previously, the code that generated eval results when S3_ENABLED=False
was separate from the code that saved a result file when
S3_ENABLED=True. So theoretically the two responses could differ, which
seems bad. Now they use the same code, which reduces duplication and
ensures the results are consistent (and tested)