Remove field_values from EvaluatorResult model
In a previous version of Metro2, we used to save a subset of data from
an AccountActivity record on each EvaluatorResult record in a field
called `field_values`. We then returned this data in the evaluator
results view endpoint.
In PR #358, we updated the evaluator code to not save field_values
anymore. Instead, in order to return specific AccountActivity data
via the evaluator results view endpoint, we join the EvaluatorResults
table to the AccountActivity table.
For some reason (??), in that PR, we didn't also remove the field_values
field from the EvaluatorResult model.