Add a serializer compatible w/ evaluator results endpoint
For using our DRF paginator (and in the future, DRF filters) for evaluator results we have to be able to serialize its output. The paginator evaluates the queryset and returns a page-based encapsulation of the resulting objects.

This serializer tries to mirror the results that `upload_utils.generate_json_sample()` generates when calling `values()` on the queryset. That function provides a flat list of fields on `AccountActivity` with its one-to-one relations included using Django's double-underscore notation, e.g. `account_holder__ecoa`. Then it filters the results by the evaluator metadata's combined results summary field list, so that we only return fields that are relevant to the evaluator.

DRF has strong opinions about API and model structure that don't hold for ours, so to facilitate this the serializer does two things:
1. Iterates over reverse related models to extract fields that we want to include, and adds them to the serializer as our expected `model_name__field`
2. Filters the fields based on `evaluator.result_summary_fields()` so that only relevant fields get serialized.

I believe this effort is worthwhile, because using DRF's API for serialization enables easier use of their paginator API (and hopefully their filter API).