Add pagination to evaluator results endpoint
This change adds pagination to the evaluator results endpoint and does so by converting the view to a `ListAPIView` subclass, hopefully setting the endpoint up for using Django Rest Framework's built-in filtering.

This new view checks a "view" parameter, if that parameter is not given or is "sample" the view will return either the S3 or sample_id-filtered results that it has previously. If the parameter is "all" the view will return all matching results for the evaluator, paginated.

The paginator and serializer are used for both sets of results, the sample and all results. The pagination size is set to the configured sample size, so there should always be 1 page of results for the samples.