Allow unlocking my own account via Jenkins
As the user administrator of Metro2, I have the authority to update the
is_active flag on user records if they need to regain access or end
access to the application. I usually do that through the administrator
interface, but in the case of my own user account, if it has become
deactivated, I don't have access to the admin interface to re-activate
it. This management command fills that gap.