Make the list of evaluators configurable
This change adds a `METRO2_EVALUATORS` setting to the `base.py` settings file, and modifies the `Evaluate` class to load the list of evaluator ID and callable from that dictionary of ID string and dotted-import string.

This should enable us to configure evaluators that are provided by other Python packages, and thus move our existing evaluators, without impacting the ability of the application to load and run them.

Because the `METRO2_EVALUATORS` is the same effective format (just with an import) as the `evaluators` dict in each separate evaluator file, I have removed those dicts.